Split out RawTableMasterIntegrationTest
Format
